requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
transformers==4.57.1
sentence-transformers==2.7.0
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.26.4
requests==2.32.3
google-generativeai==0.8.2
selenium==4.38.0
python-dotenv==1.0.1         # for OpenAI rerank (optional)
